Move smoldot mocking helpers into a separate package
