+++ packages/smoldot-test-utils/.gitignore
@@ -1,0 +1,11 @@
+/dist
+
+# dependencies
+/node_modules
+
+# misc
+.DS_Store
+yarn-error.log*
+.nyc_output/
+.chains/
+tsconfig.tsbuildinfo
