+++ packages/smoldot-test-utils/.npmignore
@@ -1,0 +1,4 @@
+tsconfig.json
+src/
+src/examples/
+dist/**/*.test.js
