+++ packages/smoldot-test-utils/src/index.ts
@@ -1,0 +1,116 @@
+import sinon from 'sinon';
+import { SmoldotClient, SmoldotOptions } from 'smoldot';
+
+type RpcResponder = (request: string) => string;
+
+const systemHealthReponse = (id: number, peerCount: number) => {
+  return `{"jsonrpc":"2.0","id":${id} ,"result":{"isSyncing":true,"peers":${peerCount},"shouldHavePeers":true}}`;
+}
+
+// Always responds to a health request with a response that has peers
+const healthyResponder = (requestJSON: string) => {
+  const id = JSON.parse(requestJSON).id;
+  return systemHealthReponse(id, 1);
+}
+
+// Always responds to a health request with a response that has no peers
+const unhealthyResponder = (requestJSON: string) => {
+  const id = JSON.parse(requestJSON).id;
+  return systemHealthReponse(id, 0);
+}
+
+// Always responds with an error reponse
+export const erroringResponder = (requestJSON: string) => {
+  const id = JSON.parse(requestJSON).id;
+  return `{ "id": ${id}, "jsonrpc": "2.0", "error": {"code": 666, "message": "boom!" } }`;
+}
+
+// Orchestrates a sequence of has peers / has no peers responses
+export const customHealthResponder = (hasPeers: boolean[]) => {
+  return (requestJSON: string) => {
+    if (hasPeers.shift()) {
+      return healthyResponder(requestJSON);
+    } else {
+      return unhealthyResponder(requestJSON);
+    }
+  };
+}
+
+// dev chains never have peers
+const devChainHealthResponse = (id: number) => {
+  return `{"jsonrpc":"2.0","id":${id} ,"result":{"isSyncing":true,"peers":0,"shouldHavePeers":false}}`;
+}
+
+export const devChainHealthResponder = (requestJSON: string) => {
+  const id = JSON.parse(requestJSON).id;
+  return devChainHealthResponse(id);
+
+}
+
+// Mimics the behaviour of the WASM light client by deferring a call to 
+// `json_rpc_callback` after it is called which returns the response
+// returned by the supplied `responder`.
+//
+// If this is called with an rpc string containing the word subscribe (but not
+// unsubscribe), it is considered a subscription and `responder` must have 2
+// mock responses for each subscriptionrequest that the test will make.  The
+// first returning the subscription id.  The second a response to the
+// subscription.
+const fakeRpcSend = (options: SmoldotOptions, responder: RpcResponder, healthResponder: RpcResponder) => {
+  return (rpcRequest: string) => {
+    process.nextTick(() => {
+      if (/system_health/.test(rpcRequest)) {
+        options.json_rpc_callback(healthResponder(rpcRequest));
+        return;
+      }
+
+      // non-health reponse
+      options.json_rpc_callback(responder(rpcRequest))
+      if (/(?<!un)[sS]ubscribe/.test(rpcRequest)) {
+        options.json_rpc_callback(responder(rpcRequest))
+      }
+    });
+  };
+}
+
+// Creates a fake smoldot. Calls to send use `fakeRpcSend` to mimic the light
+// client behaviour
+export const mockSmoldot = (responder: RpcResponder, healthResponder = healthyResponder) => {
+  return {
+    start: async (options: SmoldotOptions): Promise<SmoldotClient> => {
+      return Promise.resolve({
+        // fake the async reply by using the reponder to format
+        // a reply via options.json_rpc_callback
+        send_json_rpc: fakeRpcSend(options, responder, healthResponder)
+      });
+    }
+  };
+};
+
+// Creates a spying `smoldot` that records calls to `json_rpc_send` in `rpcSpy`
+// and then uses `fakeRpcSend` to mimic the light client behaviour.
+export const smoldotSpy = (responder: RpcResponder, rpcSpy: any, healthResponder = healthyResponder) => {
+  return {
+    start: async (options: SmoldotOptions): Promise<SmoldotClient> => {
+      return Promise.resolve({
+        send_json_rpc: (rpc: string) => {
+          // record the message call
+          rpcSpy(rpc);
+          // fake the async reply using the responder
+          fakeRpcSend(options, responder, healthResponder)(rpc);
+        }
+      });
+    }
+  };
+};
+
+// Creates an `RpcResponder` with a queue of mock reponses to respond with.
+export const respondWith = (jsonResponses: string[]) => {
+  return (_: string) => {
+    const mockResponse = jsonResponses.shift();
+    if (!mockResponse) { 
+      throw new Error("json_rpc_callback was called but there are no mock responses left to respond with"); 
+    }
+    return mockResponse;
+  };
+}
